ambil dari main

+++ deluser.php
@@ -8,15 +8,21 @@ $data = json_decode($json);
 
 $id = $data->id;
 
+$user_data = $user->get_data(Session::get('username'));
 $SadminUser = new ControllerSuperadminUser();
-$img = $SadminUser->fetch_img($id);
 
-if ($SadminUser->delete_admin($id)) {
-    $datauser = $SadminUser->delete_admin($id);
-    if (file_exists($img)) {
-        unlink($img);
-    }
-    echo json_encode(['status' => 'success']);
-} else {
+if ($user_data['id_admin'] == $id) {
     echo json_encode(['status' => 'failed']);
+} else {
+    $img = $SadminUser->fetch_img($id);
+
+    if ($SadminUser->delete_admin($id)) {
+        $datauser = $SadminUser->delete_admin($id);
+        if (file_exists($img)) {
+            unlink($img);
+        }
+        echo json_encode(['status' => 'success']);
+    } else {
+        echo json_encode(['status' => 'failed']);
+    }
 }
